+++ bootstrap/src/main/java/onlydust/com/marketplace/api/bootstrap/configuration/DomainConfiguration.java
@@ -19,6 +19,7 @@ import onlydust.com.marketplace.api.domain.service.*;
 import onlydust.com.marketplace.api.postgres.adapter.PostgresGithubAdapter;
 import onlydust.com.marketplace.api.postgres.adapter.PostgresProjectAdapter;
 import onlydust.com.marketplace.api.postgres.adapter.PostgresUserAdapter;
+import onlydust.com.marketplace.kernel.port.output.ImageStoragePort;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.retry.annotation.EnableRetry;
