+++ infrastructure/aws-s3-adapter/src/main/java/onlydust/com/marketplace/api/aws/s3/adapter/AwsS3Adapter.java
@@ -7,7 +7,7 @@ import com.amazonaws.services.s3.model.PutObjectResult;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import onlydust.com.marketplace.kernel.exception.OnlyDustException;
-import onlydust.com.marketplace.api.domain.port.output.ImageStoragePort;
+import onlydust.com.marketplace.kernel.port.output.ImageStoragePort;
 import org.apache.commons.codec.digest.DigestUtils;
 
 import javax.imageio.ImageIO;
